fix(hogql): type date subtraction as integer, cast float durations
Addresses two regressions found in review.

`Date - Date` is an Int32 count of days in ClickHouse, and an integer
unifies with a Decimal branch to a Decimal the way ClickHouse does. Typing
that duration as Float made `if(cond, date - date, <decimal>)` unify to
Float, so the printer skipped divideDecimal and the division failed with
code 69, "Decimal result's scale is less than argument's one".

A DateTime duration stays Float, because HogQL carries no datetime
precision and the runtime type is Int32 for DateTime and Decimal(18, 6)
for DateTime64. `toDate` and `toDateTime` reject a Decimal argument with
code 44, so the ClickHouse printer now casts a Float-typed argument with
toFloat64. That also makes `toDate(timestamp - timestamp)` work on the
events table, where the DateTime64 duration failed before this change.

`toDate` also lists IntegerType among its overloads now, so a day count
selects the plain constructor like `toDateTime` already does, instead of
toDateOrNull, which returns Nullable and rejects a Float with code 43.

Verified against ClickHouse 26.7: Date - Date and DateTime - DateTime are
Int32, DateTime64 - DateTime64 is Decimal(18, 6), divideDecimal requires
both arguments Decimal, toDate and toDateTime reject a Decimal, and both
accept the toFloat64 form, including a nullable one.

Ran ruff check and ruff format directly, because the pre-commit hook needs
a venv this sandbox cannot build (uv has no Python 3.13.13 available).

+++ posthog/hogql/test/test_resolver.py
@@ -1461,10 +1461,11 @@ class TestResolver(BaseTest):
         node = cast(ast.SelectQuery, resolve_types(node, self.context, dialect="clickhouse"))
         self._assert_first_columm_is_type(node, ast.FloatType(nullable=False))
 
-        # dates subtract to a duration as well, so this must not fall through to UnknownType
+        # dates subtract to an Int32 count of days, so this must neither fall through to UnknownType
+        # nor claim Float, which loses divideDecimal on a division of a decimal branch
         node = self._select("select toDate(timestamp) - toDate(timestamp) as key from events")
         node = cast(ast.SelectQuery, resolve_types(node, self.context, dialect="clickhouse"))
-        self._assert_first_columm_is_type(node, ast.FloatType(nullable=False))
+        self._assert_first_columm_is_type(node, ast.IntegerType(nullable=False))
 
         # timestamp shifted by an integer stays a datetime
         node = self._select("select timestamp - 1 as key from events")
